requirements.txt
numpy>=1.16
pytest>=2.6
scipy>=1.4
matplotlib
importlib_resources
sympy
numpydoc
sphinx_copybutton
sphinx-autoapi
nbsphinx
sphinx_rtd_theme
nbsphinx-link
docutils >= 0.18  # Needed to show bullet points in sphinx_rtd_theme

# Additional dependencies for electron density analysis
pandas
seaborn
qc-iodata
qc-gbasis
